lambdaで実行するのに加えた修正をコミット
