+++ requirements.txt
@@ -1,33 +1,37 @@
-aiohappyeyeballs==2.6.1
-aiohttp==3.12.15
-aiosignal==1.4.0
-annotated-types==0.7.0
-anyio==4.10.0
+slack_sdk
 arxiv==2.2.0
-async-timeout==5.0.1
-attrs==25.3.0
-certifi==2025.8.3
-charset-normalizer==3.4.3
-distro==1.9.0
-feedparser==6.0.11
-frozenlist==1.7.0
-h11==0.16.0
-httpcore==1.0.9
-httpx==0.28.1
-idna==3.10
-jiter==0.10.0
-multidict==6.6.4
-openai==1.99.9
-propcache==0.3.2
-pydantic==2.11.7
-pydantic-core==2.33.2
-requests==2.32.4
-sgmllib3k==1.0.0
-slack==0.0.2
-slack-sdk==3.36.0
-sniffio==1.3.1
-tqdm==4.67.1
-typing-extensions==4.14.1
-typing-inspection==0.4.1
-urllib3==2.5.0
-yarl==1.20.1
+openai
+feedparser>=6.0.0
+# aiohappyeyeballs==2.6.1
+# aiohttp==3.12.15
+# aiosignal==1.4.0
+# annotated-types==0.7.0
+# anyio==4.10.0
+# arxiv
+# async-timeout==5.0.1
+# attrs==25.3.0
+# certifi==2025.8.3
+# charset-normalizer==3.4.3
+# distro==1.9.0
+# feedparser
+# frozenlist==1.7.0
+# h11==0.16.0
+# httpcore==1.0.9
+# httpx==0.28.1
+# idna==3.10
+# jiter==0.10.0
+# multidict==6.6.4
+# openai==1.99.9
+# propcache==0.3.2
+# pydantic==2.11.7
+# pydantic-core==2.33.2
+# requests==2.32.4
+# # sgmllib3k
+# # slack
+# slack-sdk
+# sniffio==1.3.1
+# tqdm==4.67.1
+# typing-extensions==4.14.1
+# typing-inspection==0.4.1
+# urllib3==2.5.0
+# yarl==1.20.1
